use 32 bit indices

+++ ios/graphics/Model/Icosahedron/Icosahedron.swift
@@ -73,7 +73,7 @@ final class Icosahedron: BaseGraphicsObject {
 
         encoder.drawIndexedPrimitives(type: .triangle,
                                       indexCount: indicesCount,
-                                      indexType: .uint16,
+                                      indexType: .uint32,
                                       indexBuffer: indicesBuffer,
                                       indexBufferOffset: 0)
     }
